search background by id done

+++ src/main/java/al/cherbourg_photographers/repository/WorkshopBackgroundEntityRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface WorkshopBackgroundEntityRepository extends CrudRepository<WorkshopBackgroundEntity, Long> {
+public interface WorkshopBackgroundEntityRepository extends CrudRepository<WorkshopBackgroundEntity, Integer> {
 }
